+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/ngneat/loadoff/compare/v1.1.0...v1.2.0) (2021-08-01)
+
+
+### Features
+
+* 🎸 createAsyncStore ([735313b](https://github.com/ngneat/loadoff/commit/735313bc08e6f456b806bc2e8ae69d7d1588f1fa))
+* Add missing generic Error type to toAsyncState functions ([c237e3c](https://github.com/ngneat/loadoff/commit/c237e3c364a9ae09d1004c851cc7f9d9bd1d1133))
+
+
+### Bug Fixes
+
+* add default empty value for state parameter in createAsyncState function ([48296d5](https://github.com/ngneat/loadoff/commit/48296d5cfcd908db5e4f4efb7cda77f57ae1f3a2))
+* Add missing @ngneat/subscribe dependency, and update pre-commit to be able to run lint-staged ([531202d](https://github.com/ngneat/loadoff/commit/531202daa1eb244940eb50d5ad338fdcf1eb79c0))
+* Make e2e green ([c14e93a](https://github.com/ngneat/loadoff/commit/c14e93ab6f7b17349f13afd11e263673df240340))
+
 ## [1.1.0](https://github.com/ngneat/loadoff/compare/v1.0.0...v1.1.0) (2021-06-16)